Added initialize and getAnomalyScore methods to the OneClassClassifier interface.
The initialize method allows a one class classifier to be initialized with a starting set of training instances.
The getAnomalyScore method is used when an anomaly score is needed instead of a vote. The higher an instance's anomaly score is, the more likely it is an anomaly.

These methods were implemented in Autoencoder, HSTrees and NearestNeighbourDescription. All three also had minor code clean ups.